fix: signer page ssr error
